線分に点が重複する不具合を修正

+++ WinOGL/WinOGL/CMath.h
@@ -24,7 +24,7 @@ public:
     /// @brief 三角形の面積を計算する．
     /// @param triangle 三角形
     /// @return 面積
-    static float TriangleArea (CShape* triangle);
+    static double TriangleArea (CShape* triangle);
 
     /// @brief 点の垂線と直線間の交点を計算する．
     /// @param p      頂点
@@ -111,7 +111,7 @@ public:
     /// @param p_b_s 2 つ目のベクトルの始点
     /// @param p_b_e 2 つ目のベクトルの終点
     /// @return 2 つのベクトルの外積
-    static float Outer2DSize (CVertex* p_a_s, CVertex* p_a_e, CVertex* p_b_s, CVertex* p_b_e);
+    static double Outer2DSize (CVertex* p_a_s, CVertex* p_a_e, CVertex* p_b_s, CVertex* p_b_e);
 
     /// @brief 2 つのベクトルのなす面の法線を計算する．
     /// @param p_a_s   1 つ目のベクトルの始点
